+++ Collibri/ClientApp/src/components/JoinRoom.js
@@ -1,0 +1,15 @@
+import React from "react";
+import './LandingPage.css';
+
+export const JoinRoom = () => {
+    const handleJoinRoom = () => {
+        // Implement join button
+        alert("Work in progress");
+    }
+
+    return (
+        <div>
+            <button className={"btn buttons"} onClick={handleJoinRoom}>Join Room</button>
+        </div>
+    );
+}
